Move load_api_spec to the API module
